+++ lib/urlwatch/reporters.py
@@ -630,31 +630,38 @@ class SlackReporter(TextReporter):
         text = '\n'.join(super().submit())
 
         if not text:
-            logger.debug('Not calling slack API (no changes)')
+            logger.debug('Not calling {} API (no changes)'.format(self.__kind__))
             return
 
         result = None
         for chunk in chunkstring(text, self.max_length, numbering=True):
-            res = self.submit_to_slack(webhook_url, chunk)
+            res = self.submit_chunk(webhook_url, chunk)
             if res.status_code != requests.codes.ok or res is None:
                 result = res
 
         return result
 
-    def submit_to_slack(self, webhook_url, text):
-        logger.debug("Sending slack request with text:{0}".format(text))
+    def submit_chunk(self, webhook_url, text):
+        logger.debug("Sending {} request with text: {}".format(self.__kind__, text))
         post_data = {"text": text}
         result = requests.post(webhook_url, json=post_data)
         try:
             if result.status_code == requests.codes.ok:
-                logger.info("Slack response: ok")
+                logger.info("{} response: ok".format(self.__kind__))
             else:
-                logger.error("Slack error: {0}".format(result.text))
+                logger.error("{} error: {}".format(self.__kind__, result.text))
         except ValueError:
             logger.error(
-                "Failed to parse slack response. HTTP status code: {0}, content: {1}".format(result.status_code,
-                                                                                             result.content))
+                "Failed to parse {} response. HTTP status code: {}, content: {}".format(self.__kind__,
+                                                                                        result.status_code,
+                                                                                        result.content))
         return result
+
+
+class MattermostReporter(SlackReporter):
+    """Send a message to a Mattermost channel"""
+
+    __kind__ = 'mattermost'
 
 
 class DiscordReporter(TextReporter):